consider remote config
